Update repository name
Before tagging a release, I realized we need to update the permissions used by GitHub Actions for deploying into the AWS account.

Instead of API keys (which can leak), we are using GitHub OIDC to authenticate into the AWS account. The way this works is that a "stack" is deployed into the AWS account and it creates a role that allows this repository (via its name).

Since the repository name changed, you will need to update the role by updating the stack in CloudFormation.

I am not 100% sure if you will need to re-upload the YAML template, or if you'll be able to change the parameter and redeploy straightaway. In the extreme case where it's confusing, you can delete the stack and re-create it following the instructions listed in the README.md that I just edited.

I can help if you have any trouble.